added user auth with mongo

+++ app/Screens/MessagesScreen.js
@@ -1,50 +1,69 @@
-import React from 'react';
-import { View, Text, FlatList } from 'react-native';
+import React,  {useEffect, useContext } from 'react';
+import  { useState } from 'react';
+import { FlatList  ,StyleSheet, View, Alert } from 'react-native';
 
 import Swipeable from 'react-native-gesture-handler/Swipeable';
 
+import IconComponent from '../components/AppComponents/IconComponent';
 import ListComponent from '../components/ListComponents/ListComponent';
-import Screen from '../components/ScreenComponents/Screen';
 import ListItemSeperator from '../components/ListComponents/ListItemSeperator';
 import ListItemDeleteAction from '../components/ListComponents/ListItemDeleteAction';
-import { useState } from 'react';
 
-const initMessages = [
-    {
-        id:1,
-        title:'T1',
-        description: 'D1',
-        image:require('../assets/profileImage.png')
-    },
-    {
-        id:2,
-        title:'T2',
-        description: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industrys standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.',
-        image:require('../assets/profileImage.png')
-    },
+import colors from '../configs/colors';
 
-];
+import messagesApi from '../api/messages';
 
-function MessagesScreen(props) {
-    const [messages, setMessages] = useState(initMessages);
+import routes from '../Navigators/routes';
+import AuthContext from '../auth/context';
+
+
+
+function MessagesScreen({navigation}) {
     const [refreshing, setRefreshing] = useState(false);
-    
+    const getMessagesApi = useApi(messagesApi.getMessages);
+    const {user} = useContext(AuthContext);
+    // const [messages, setMessages] = useState();
+    useEffect(() => {
+        getMessagesApi.request(user);
+     }, [])
+     
+    // setMessages(getMessagesApi.data);
+
+     const refresh =() =>{
+         setRefreshing(true);
+         getMessagesApi.request(user);
+         setRefreshing(false);
+    }
+
     const handleDelete = message =>{
         setMessages(messages.filter((m) => m.id !== message.id));
     };
+
+    const handleNoData = () => {
+        navigation.navigate(routes.ACCOUNT);
+    }
+    if ((getMessagesApi.data) &&  (getMessagesApi.data.length === 0)){
+        Alert.alert(`Hi, ${user.name}`,'Your message box is empty',[{text:'OK', onPress:handleNoData}])
+        return <View></View>
+    }
     return (
-        <Screen>
+        <View style={styles.container}>
             <FlatList
-                data={messages}
-                keyExtractor={(message)=>message.id.toString()}
+                data={getMessagesApi.data}
+                keyExtractor={(message)=>message._id}
                 renderItem={({item})=>(
                     <Swipeable renderRightActions={()=> 
                         <ListItemDeleteAction 
                             onPress={()=>handleDelete(item)}/>}>
                         <ListComponent
-                            title={item.title}
-                            subtitle = {item.description}
-                            image={item.image}
+                            title={item.fromUser.name}
+                            subtitle = {item.content}
+                            image={{uri:item.fromUser.images[0].url}}
+                            ImageComponent={<IconComponent 
+                                                    name="account"
+                                                    size={40}
+                                                    color={colors.darkBackground} 
+                                                />}
                             onPress={()=> console.log(item)}
                             showCheveron={true}                        
                             />
@@ -52,18 +71,19 @@ function MessagesScreen(props) {
                 )}
                 ItemSeparatorComponent={ListItemSeperator}
                 refreshing={refreshing}
-                onRefresh={()=>{
-                    setMessages([
-                        {
-                            id:3,
-                            title:'T3',
-                            description: 'D3',
-                            image:require('../assets/mosh.jpg')
-                        },
-                    ]);
-                }}
+                onRefresh={refresh}
             />
-        </Screen>
+        </View>
     )
+    // }
 }
+
+const styles = StyleSheet.create({
+    container: {
+        justifyContent: "center",
+        alignItems: "center",
+        marginTop:40,
+        flex:1,
+    },
+});
 export default MessagesScreen;
